fix: remove Pro 2 from USB connect device label
